rename configuration fields tab component files
