+++ src/angular-app/languageforge/lexicon/settings/configuration/configuration-fields.component.html
